creditandsupport updated
new higher position on desktop as people don't seem to scroll that much. hope it doesn't break too much the flow of the page, I've been reluctant to do this for years but I think that with this new copywriting it's informative and adds value to the whole. Plus, if it can drive more donations it would be great for the collective.